requirements.txt
flask==3.0.3
python-dotenv==1.0.1
openpyxl==3.1.5
pandas==2.2.2
dnspython==2.6.1
